+++ app/page.tsx
@@ -4,6 +4,7 @@ import AllIssuesStatus from '@/components/AllIssuesStatus';
 import IssueChart from '@/components/IssueChart';
 import LatestIssues from '@/components/LatestIssues';
 import { Flex, Grid } from '@radix-ui/themes';
+import { Metadata } from 'next';
 
 export default function Home() {
   return (
@@ -16,3 +17,9 @@ export default function Home() {
     </Grid>
   );
 }
+
+
+export const metadata: Metadata = {
+  title: 'Tracker.io - Dashboard',
+  description: 'View all project issues',
+};